Sonic PM Geo Guide Book (v1.0)
This is the geography guide book for the Sonic the Hedgehog Protamedra fan project! This guide book has the full list of zones, three easy to read maps, and a few other nifty features packed into it! Make sure to have this on you while reading through other parts of the project, that way you can see where everything is located!