Sonic PM Overview Book (v1.0)
This is the overview book for the Sonic the Hedgehog Protamedra fan project! This guide book has information concerning the history of the project, a list of series events, hints, secrets, interview material, and so much more! Make sure to look through this before tackling any of the episode documents!